DBZ-3668 Extracting SPI package
