+++ debezium-connector-postgres/src/main/java/io/debezium/connector/postgresql/converters/PostgresCloudEventsProvider.java
@@ -9,10 +9,10 @@ import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.Struct;
 
 import io.debezium.connector.postgresql.Module;
-import io.debezium.converters.CloudEventsMaker;
-import io.debezium.converters.CloudEventsProvider;
-import io.debezium.converters.RecordParser;
-import io.debezium.converters.SerializerType;
+import io.debezium.converters.spi.CloudEventsMaker;
+import io.debezium.converters.spi.CloudEventsProvider;
+import io.debezium.converters.spi.RecordParser;
+import io.debezium.converters.spi.SerializerType;
 
 /**
  * An implementation of {@link CloudEventsProvider} for PostgreSQL.
