+++ debezium-connector-postgres/src/main/java/io/debezium/connector/postgresql/converters/PostgresRecordParser.java
@@ -11,7 +11,7 @@ import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.Struct;
 import org.apache.kafka.connect.errors.DataException;
 
-import io.debezium.converters.RecordParser;
+import io.debezium.converters.spi.RecordParser;
 import io.debezium.data.Envelope;
 import io.debezium.util.Collect;
