+++ debezium-core/src/main/java/io/debezium/converters/spi/CloudEventsProvider.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.converters;
+package io.debezium.converters.spi;
 
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.Struct;
